+++ NuGet/System.Linq.Dynamic.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
     <metadata>
         <id>System.Linq.Dynamic.Library</id>
-        <version>1.1.14</version>
+        <version>1.1.15</version>
         <title>Dynamic Linq Library</title>
         <authors>Nathan Arnott,Microsoft</authors>
         <owners>Nathan Arnott</owners>
@@ -13,18 +13,42 @@
 
 This package is more than a simple NuGet deployment of the Dynamic Linq Library. It is a living project, and contains XML documentation with examples.</description>
         <summary>Supports Framework 3.5 and 4.0+</summary>
-        <releaseNotes>1.1.14 Added support for Silverlight 5. A special thanks to jkuehner for providing this.
-		
+        <releaseNotes>1.1.15 Several changes by meatGUY. 
+        
+Support for Text Aliases for the Logical and Relation operations
+  "eq" for ==
+  "ne" for !=
+  .. and so on
+  
+Support for Short naming for primitive types of .NET
+  int for Int32
+  float for Single
+  .. and so on, full list can be found on MSDN (https://msdn.microsoft.com/en-us/library/ms228360(v=vs.90).aspx)
+
+Support for Equality check with DateTime object in string form:
+
+DateTime == "string datetime"
+
+Support for TryParse usage during explicit cast.
+
+Shorthand for explicit cast from string, now it can be done using T"..".
+
+Updated documentation accordingly to changes.
+
+Equality check for Guid (the explanation is placed in code as comment)
+  
+1.1.14 Added support for Silverlight 5. A special thanks to jkuehner for providing this.
+    
 Added several new helper methods, including First, FirstOrDefault, Last, LastOrDefault, Single, SingleOrDefault, and Sum.
 
 1.1.13 Added Dynamic Join Method.
-		
+    
 Added dynamic context keywords "parent" and "root", to the existing keyword "it".  Also added symbol equivalents to those keywords ($ - it, ^ - parent, and ~ - root). Added ability to GlobalConfig to disable the context keywords.
-		
+    
 1.1.12 Added dynamic method for GroupByMany. Support for logical operators. Better Guid and Enum Support.  Added support for shift operators.
-		
+    
 Added static GlobalConfig class, and IDynamicLinkCustomTypeProvider to allow developers to add custom understood types to Dynamic Linq without the need to recompile.
-		
+    
 1.1.9 Added In operator for both static inline lists, as well as a single parameter list (similiar to contains).
 
 1.1.8 Added support for the use of Single(), SingleOrDefault(), First(), FirstOrDefault(), OrderBy(), OrderByDescending(), and Contains() in string expressions.
