+++ Src/System.Linq.Dynamic.dnx/ClassFactory.cs
