+++ Src/System.Linq.Dynamic.dnx/DynamicClass.cs
