+++ Src/System.Linq.Dynamic.dnx/DynamicLinqTypeProvider.cs
@@ -29,7 +29,7 @@ namespace System.Linq.Dynamic
 
         static IEnumerable<Type> FindTypesMarkedWithAttribute()
         {
-#if !(NETFX_CORE || DNXCORE50)
+#if !(NETFX_CORE || DNXCORE50 || DOTNET5_4)
             return AppDomain.CurrentDomain.GetAssemblies()
 #if !(NET35)
                 .Where(x => !x.IsDynamic)
