+++ Src/System.Linq.Dynamic.dnx/DynamicObjectClass.cs
