+++ Src/System.Linq.Dynamic.dnx/DynamicProperty.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace System.Linq.Dynamic
+{
+    internal class DynamicProperty
+    {
+        string name;
+        Type type;
+
+        public DynamicProperty(string name, Type type)
+        {
+            this.name = name;
+            this.type = type;
+        }
+
+        public string Name
+        {
+            get { return name; }
+        }
+
+        public Type Type
+        {
+            get { return type; }
+        }
+    }
+}
