+++ Src/System.Linq.Dynamic.dnx/GlobalConfig.cs
