+++ Src/System.Linq.Dynamic.dnx/GroupResult.cs
