+++ Src/System.Linq.Dynamic.dnx/ParseException.cs
@@ -8,7 +8,7 @@ namespace System.Linq.Dynamic
     /// Represents errors that occur while parsing dynamic linq string expressions.
     /// </summary>
     [SuppressMessage("Microsoft.Design", "CA1032:ImplementStandardExceptionConstructors")]
-#if !(SILVERLIGHT || DNXCORE50)
+#if !(SILVERLIGHT || DNXCORE50 || DOTNET5_4)
     [Serializable]
 #endif
     public sealed class ParseException : Exception
@@ -43,7 +43,7 @@ namespace System.Linq.Dynamic
             return string.Format(CultureInfo.CurrentCulture, Res.ParseExceptionFormat, Message, _position);
         }
 
-#if !(SILVERLIGHT || DNXCORE50)
+#if !(SILVERLIGHT || DNXCORE50 || DOTNET5_4)
         ParseException(SerializationInfo info, StreamingContext context)
             : base(info, context)
         {
