+++ Src/System.Linq.Dynamic.dnx/TextFile.txt
@@ -1,0 +1,44 @@
+﻿        "net35": {
+            "frameworkAssemblies": {
+            }
+        },
+        "net40": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "net45": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "net451": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "net452": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "net46": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "net461": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "dnx451": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
+        "dnx452": {
+            "frameworkAssemblies": {
+                "System.Data.Entity": "4.0.0.0"
+            }
+        },
