--- Src/System.Linq.Dynamic/DynamicOrdering.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
-
-namespace System.Linq.Dynamic
-{
-    internal class DynamicOrdering
-    {
-        public Expression Selector;
-        public bool Ascending;
-    }
-}
