+++ Src/System.Linq.Dynamic/System.Linq.Dynamic.csproj
@@ -61,29 +61,67 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClassFactory.cs" />
-    <Compile Include="DynamicClass.cs" />
-    <Compile Include="DynamicExpression.cs" />
-    <Compile Include="DynamicLinqTypeProvider.cs" />
-    <Compile Include="DynamicObjectClass.cs" />
-    <Compile Include="DynamicQueryable.cs" />
-    <Compile Include="DynamicOrdering.cs" />
-    <Compile Include="DynamicProperty.cs" />
-    <Compile Include="ExpressionConverter.cs" />
-    <Compile Include="ExpressionParser.NetFX_backup.cs" />
-    <Compile Include="ExpressionParser.cs" />
-    <Compile Include="GlobalConfig.cs" />
-    <Compile Include="GlobalSuppressions.cs" />
-    <Compile Include="GroupResult.cs" />
-    <Compile Include="IDynamicLinkTypeProvider.cs" />
-    <Compile Include="ParseException.cs" />
+      <Compile Include="..\System.Linq.Dynamic.dnx\ClassFactory.cs">
+      <Link>ClassFactory.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicClass.cs">
+      <Link>DynamicClass.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicExpression.cs">
+      <Link>DynamicExpression.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicLinqTypeProvider.cs">
+      <Link>DynamicLinqTypeProvider.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicObjectClass.cs">
+      <Link>DynamicObjectClass.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicQueryable.cs">
+      <Link>DynamicQueryable.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicOrdering.cs">
+      <Link>DynamicOrdering.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\DynamicProperty.cs">
+      <Link>DynamicProperty.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\ExpressionConverter.cs">
+      <Link>ExpressionConverter.cs</Link>
+    </Compile>
+	    <Compile Include="..\System.Linq.Dynamic.dnx\ExpressionParser.NetFX_backup.cs">
+      <Link>ExpressionParser.NetFX_backup.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\ExpressionParser.cs">
+      <Link>ExpressionParser.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\GlobalConfig.cs">
+      <Link>GlobalConfig.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\GlobalSuppressions.cs">
+      <Link>GlobalSuppressions.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\GroupResult.cs">
+      <Link>GroupResult.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\IDynamicLinkTypeProvider.cs">
+      <Link>IDynamicLinkTypeProvider.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\ParseException.cs">
+      <Link>ParseException.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\BasicQueryable.cs">
+      <Link>BasicQueryable.cs</Link>
+    </Compile>
+	<Compile Include="..\System.Linq.Dynamic.dnx\Res.cs">
+      <Link>Res.cs</Link>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="BasicQueryable.cs" />
-    <Compile Include="Res.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
     <None Include="Properties\key.snk" />
+  </ItemGroup>
+  <ItemGroup>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
